feat: move util/nric-validation to root shared folder
